requirements.txt
fastapi
uvicorn
slowapi
pydantic
pydantic-settings
python-dotenv
PyJWT
cryptography
bcrypt
python-multipart
langchain
langchain-community
langchain-classic
langchain-core
langchain-text-splitters
langchain-google-genai
langchain-mistralai
langsmith
chromadb
pypdf
PyPDF2
docx2txt
pdf2image
pytesseract
pillow
numpy
pandas
scikit-learn
scipy
tiktoken
tokenizers
httpx
requests
PyYAML
tenacity
annotated-types
typing-extensions
orjson
anyio
google-auth
googleapis-common-protos
sqlalchemy
psycopg2-binary
email-validator

